Implement user service with RBAC and JDBC-based auth server authentication
- Add user-service with Liquibase migrations for RBAC schema (users, roles, permissions)
- Add JdbcUserDetailsService to auth-server for database-backed authentication
- Register k8s-dashboard-client OAuth2 PKCE client on auth-server
- Fix user-service Dockerfile to use Java 24 runtime matching compile target

+++ src/backend/core/user-service/src/main/java/dev/agasen/core/user/persistence/PermissionRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Collection;
 import java.util.stream.Collectors;
 
-public interface PermissionRepository extends JpaRepository< Permission, Integer > {
+public interface PermissionRepository extends JpaRepository< Permission, Long > {
 
    Collection< Permission > findByNameIn( Collection< String > names );
